Centralize tests for notifications
